Extract record sport resolver from page export
